Grape Leaf Disease Detection.js
This method can be used for Supervised (Random Forest) Classification of Grape Fields successfully. It contains algorithms of classification and gets the job done quickly. The combination of SAR and MSI data is a good selection for crop classification. 
Further the grape disease detection is done by the application of indices and Unsupervised classification. The unsupervised classification generates 3 clusters of Grape pixels. These three clusters are distinguished as Healthy, Unhealthy and Moderately Healthy crops.